minesweeper(step1): Controller 제거
